requirements.txt
python-dotenv==1.0.1
requests==2.32.3
openai==0.27.2
slack_bolt==1.23.0
flask==2.2.5